split parser...except i messed up somewhere
