+++ src/frontend/parser.cpp
@@ -1,746 +1,13 @@
-#include <string>
-#include <utility>
-
-#include <tao/pegtl.hpp>
 #include <tao/pegtl/contrib/analyze.hpp>
-#include <tao/pegtl/contrib/raw_string.hpp>
 
-#include "frontend/ast/ast.h"
-#include "frontend/ast/expr.h"
-#include "frontend/ast/instruction.h"
 #include "frontend/parser.h"
-#include "frontend/utils/atom.h"
-#include "frontend/utils/operator.h"
-#include "frontend/utils/symbol_table.h"
-#include "frontend/utils/type.h"
-
-namespace pegtl = tao::pegtl;
-using namespace pegtl;
-
-template <typename Rule> using pegtl_try = pegtl::seq<pegtl::at<Rule>, Rule>;
+#include "frontend/parser/function.h"
 
 namespace frontend {
-    enum class TokenType {
-        NUMBER,
-        IDENTIFIER,
+    struct grammar
+        : pegtl::must<pegtl::seq<ignorable, pegtl::star<function>, ignorable>> {
     };
 
-    using Token = std::pair<std::string, TokenType>;
-
-    std::unique_ptr<SymbolTable> symbol_table;
-    std::vector<Token> parsed_tokens;
-    std::vector<std::unique_ptr<Type>> parsed_types;
-    std::vector<std::unique_ptr<ast::Expr>> parsed_exprs;
-    std::vector<std::unique_ptr<ast::Scope>> active_scopes;
-    BinaryOp last_op_equals;
-    std::vector<std::unique_ptr<ast::Expr>> active_args;
-    std::vector<ast::Parameter> active_params;
-
-    std::unique_ptr<Atom> popAtom() {
-        auto [token_val, token_type] = parsed_tokens.back();
-        parsed_tokens.pop_back();
-        if (token_type == TokenType::NUMBER) {
-            long long parsed_val = std::stoll(token_val);
-            int32_t truncated = static_cast<int32_t>(parsed_val);
-            uint64_t val =
-                static_cast<uint64_t>(static_cast<int64_t>(truncated));
-            return std::make_unique<AtomLiteral>(val);
-        } else
-            return std::make_unique<AtomIdentifier>(
-                symbol_table->addSymbol(token_val));
-    }
-
-    std::unique_ptr<AtomIdentifier> popIdentifier() {
-        auto [token_val, token_type] = parsed_tokens.back();
-        parsed_tokens.pop_back();
-        return std::make_unique<AtomIdentifier>(
-            symbol_table->addSymbol(token_val));
-    }
-
-    std::unique_ptr<ast::Expr> popExpr() {
-        auto expr = std::move(parsed_exprs.back());
-        parsed_exprs.pop_back();
-        return expr;
-    }
-
-    std::unique_ptr<Type> popType() {
-        auto type = std::move(parsed_types.back());
-        parsed_types.pop_back();
-        return type;
-    }
-
-    std::unique_ptr<ast::Instruction> popInstruction() {
-        auto i = std::move(active_scopes.back()->getInstructions().back());
-        active_scopes.back()->getInstructions().pop_back();
-        return i;
-    }
-
-    std::unique_ptr<ast::Scope> popScope() {
-        auto scope = std::move(active_scopes.back());
-        active_scopes.pop_back();
-        return scope;
-    }
-
-    //
-    // Grammar
-    //
-    // Whitespace/ignorables are handled by
-    // program structure constructs.
-    //
-
-    // Tokens
-    struct keyword_return : TAO_PEGTL_STRING("return") {};
-    struct keyword_if : TAO_PEGTL_STRING("if") {};
-    struct keyword_else : TAO_PEGTL_STRING("else") {};
-    struct keyword_while : TAO_PEGTL_STRING("while") {};
-
-    struct keyword_short : TAO_PEGTL_STRING("short") {};
-    struct keyword_int : TAO_PEGTL_STRING("int") {};
-    struct keyword_long : TAO_PEGTL_STRING("long") {};
-    struct keyword_long_long : TAO_PEGTL_STRING("long long") {};
-
-    struct left_paren : pegtl::one<'('> {};
-    struct right_paren : pegtl::one<')'> {};
-    struct left_brace : pegtl::one<'{'> {};
-    struct right_brace : pegtl::one<'}'> {};
-    struct comma : pegtl::one<','> {};
-    struct semicolon : pegtl::one<';'> {};
-    struct plus : pegtl::one<'+'> {};
-    struct minus : pegtl::one<'-'> {};
-    struct asterisk : pegtl::one<'*'> {};
-    struct slash : pegtl::one<'/'> {};
-    struct ampersand : pegtl::one<'&'> {};
-    struct pipe : pegtl::one<'|'> {};
-    struct caret : pegtl::one<'^'> {};
-    struct tilde : pegtl::one<'~'> {};
-    struct bang : pegtl::one<'!'> {};
-    struct equal : pegtl::one<'='> {};
-    struct greater : pegtl::one<'>'> {};
-    struct less : pegtl::one<'<'> {};
-
-    // Ignorables
-    struct comment
-        : pegtl::disable<TAO_PEGTL_STRING("//"), pegtl::until<pegtl::eolf>> {};
-
-    struct whitespace_char
-        : pegtl::sor<pegtl::one<' '>, pegtl::one<'\t'>, pegtl::eol> {};
-
-    struct whitespace : pegtl::plus<whitespace_char> {};
-
-    struct ignorable : pegtl::star<pegtl::sor<whitespace_char, comment>> {};
-
-    // Comma separable list
-    template <typename Rule>
-    using pegtl_comma_list = pegtl::list<Rule, pegtl::seq<comma, ignorable>>;
-
-    // Expression
-    // Refer to operator precedence:
-    // https://en.cppreference.com/w/c/language/operator_precedence.html
-    struct expr;
-
-    // 0
-    struct identifier
-        : pegtl::seq<pegtl::plus<pegtl::sor<pegtl::alpha, pegtl::one<'_'>>>,
-                     pegtl::star<pegtl::sor<pegtl::alnum, pegtl::one<'_'>>>> {};
-
-    struct number : pegtl::seq<pegtl::opt<pegtl::sor<plus, minus>>,
-                               pegtl::plus<pegtl::digit>> {};
-    // 1
-    struct expr_1;
-    struct call_arg : pegtl::seq<expr, ignorable> {};
-    struct call
-        : pegtl::seq<identifier, ignorable, left_paren, ignorable,
-                     pegtl::opt<pegtl_comma_list<call_arg>>, right_paren> {};
-    struct parens
-        : pegtl::seq<left_paren, ignorable, expr, ignorable, right_paren> {};
-    struct value : pegtl::sor<identifier, number> {};
-    struct expr_1 : pegtl::sor<pegtl_try<call>, pegtl_try<parens>, value> {};
-
-    // 2
-    struct expr_2;
-    struct unary_plus : pegtl_try<pegtl::seq<plus, ignorable, expr_2>> {};
-    struct unary_minus : pegtl_try<pegtl::seq<minus, ignorable, expr_2>> {};
-    // struct unary_logical_not : pegtl::seq<bang, ignorable, expr_2> {};
-    struct unary_bitwise_not : pegtl_try<pegtl::seq<tilde, ignorable, expr_2>> {
-    };
-    struct deref : pegtl_try<pegtl::seq<asterisk, ignorable, expr_2>> {};
-    struct address : pegtl_try<pegtl::seq<ampersand, ignorable, expr_2>> {};
-    struct expr_2 : pegtl::sor<expr_1, unary_plus, unary_minus,
-                               unary_bitwise_not, deref, address> {};
-
-    // 3
-    struct expr_3;
-    struct multiply
-        : pegtl_try<pegtl::seq<ignorable, asterisk, ignorable, expr_2>> {};
-    struct divide : pegtl_try<pegtl::seq<ignorable, slash, ignorable, expr_2>> {
-    };
-    struct expr_3
-        : pegtl::seq<expr_2, pegtl::star<pegtl::sor<multiply, divide>>> {};
-
-    // 4
-    struct expr_4;
-    struct add : pegtl_try<pegtl::seq<ignorable, plus, ignorable, expr_3>> {};
-    struct subtract
-        : pegtl_try<pegtl::seq<ignorable, minus, ignorable, expr_3>> {};
-    struct expr_4 : pegtl::seq<expr_3, pegtl::star<pegtl::sor<add, subtract>>> {
-    };
-
-    // 5
-    struct expr_5;
-    struct left_shift
-        : pegtl_try<pegtl::seq<ignorable, less, less, ignorable, expr_4>> {};
-    struct right_shift
-        : pegtl_try<
-              pegtl::seq<ignorable, greater, greater, ignorable, expr_4>> {};
-    struct expr_5
-        : pegtl::seq<expr_4, pegtl::star<pegtl::sor<left_shift, right_shift>>> {
-    };
-
-    // 8
-    struct bitwise_and
-        : pegtl_try<pegtl::seq<ignorable, ampersand, ignorable, expr_5>> {};
-    struct expr_8 : pegtl::seq<expr_5, pegtl::star<bitwise_and>> {};
-
-    // 9
-    struct bitwise_xor
-        : pegtl_try<pegtl::seq<ignorable, caret, ignorable, expr_8>> {};
-    struct expr_9 : pegtl::seq<expr_8, pegtl::star<bitwise_xor>> {};
-
-    // 10
-    struct bitwise_or
-        : pegtl_try<pegtl::seq<ignorable, pipe, ignorable, expr_8>> {};
-    struct expr_10 : pegtl::seq<expr_9, pegtl::star<bitwise_or>> {};
-
-    struct expr : expr_10 {};
-
-    // Types
-    struct type;
-    struct type_int : keyword_int {};
-    struct type_ptr : pegtl::seq<ignorable, asterisk> {};
-    struct type : pegtl::seq<type_int, pegtl::star<type_ptr>> {};
-
-    // Op equals
-    struct add_equals : pegtl::seq<plus, equal> {};
-    struct sub_equals : pegtl::seq<minus, equal> {};
-    struct mul_equals : pegtl::seq<asterisk, equal> {};
-    struct div_equals : pegtl::seq<slash, equal> {};
-    struct lshift_equals : pegtl::seq<less, less, equal> {};
-    struct rshift_equals : pegtl::seq<greater, greater, equal> {};
-    struct and_equals : pegtl::seq<ampersand, equal> {};
-    struct or_equals : pegtl::seq<pipe, equal> {};
-    struct xor_equals : pegtl::seq<caret, equal> {};
-    struct op_equals : pegtl::sor<add_equals, sub_equals, mul_equals,
-                                  div_equals, lshift_equals, rshift_equals,
-                                  and_equals, or_equals, xor_equals> {};
-
-    // Instructions
-    struct instruction_any;
-    struct scope;
-
-    struct instruction_expr : pegtl::seq<expr, ignorable, semicolon> {};
-
-    struct instruction_declaration
-        : pegtl::seq<type, ignorable, identifier, ignorable, semicolon> {};
-
-    struct instruction_declaration_assign
-        : pegtl::seq<type, ignorable, identifier, ignorable, equal, ignorable,
-                     expr, ignorable, semicolon> {};
-
-    struct instruction_assign
-        : pegtl::seq<identifier, ignorable, equal, ignorable, expr, ignorable,
-                     semicolon> {};
-
-    struct instruction_op_assign
-        : pegtl::seq<identifier, ignorable, op_equals, ignorable, expr,
-                     ignorable, semicolon> {};
-
-    struct instruction_return
-        : pegtl::seq<keyword_return, ignorable, expr, ignorable, semicolon> {};
-
-    struct instruction_if
-        : pegtl::seq<keyword_if, ignorable, left_paren, ignorable, expr,
-                     ignorable, right_paren, ignorable, instruction_any> {};
-
-    struct instruction_if_else
-        : pegtl::seq<keyword_if, ignorable, left_paren, ignorable, expr,
-                     ignorable, right_paren, ignorable, instruction_any,
-                     ignorable, keyword_else, ignorable, instruction_any> {};
-
-    struct instruction_while
-        : pegtl::seq<keyword_while, ignorable, left_paren, ignorable, expr,
-                     ignorable, right_paren, ignorable, instruction_any> {};
-
-    struct instruction_any
-        : pegtl::sor<pegtl_try<scope>, pegtl_try<instruction_while>,
-                     pegtl_try<instruction_if_else>, pegtl_try<instruction_if>,
-                     pegtl_try<instruction_return>,
-                     pegtl_try<instruction_declaration_assign>,
-                     pegtl_try<instruction_declaration>,
-                     pegtl_try<instruction_op_assign>,
-                     pegtl_try<instruction_assign>,
-                     pegtl_try<instruction_expr>> {};
-
-    // Program structure
-    struct scope
-        : pegtl::seq<left_brace, ignorable,
-                     pegtl::star<pegtl::seq<instruction_any, ignorable>>,
-                     right_brace> {};
-
-    struct function_param : pegtl::seq<type, ignorable, identifier, ignorable> {
-    };
-    struct function
-        : pegtl::seq<type, ignorable, identifier, ignorable, left_paren,
-                     ignorable, pegtl::opt<pegtl_comma_list<function_param>>,
-                     right_paren, ignorable, scope, ignorable> {};
-
-    struct entry_point
-        : pegtl::seq<ignorable, pegtl::star<function>, ignorable> {};
-
-    struct grammar : pegtl::must<entry_point> {};
-
-    //
-    // Action
-    //
-    template <typename Rule> struct action {};
-
-    // Value token actions
-    template <> struct action<number> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            parsed_tokens.push_back({in.string(), TokenType::NUMBER});
-        }
-    };
-
-    template <> struct action<identifier> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            parsed_tokens.push_back({in.string(), TokenType::IDENTIFIER});
-        }
-    };
-
-    // Expr actions
-    template <> struct action<value> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto value = popAtom();
-            auto expr = std::make_unique<ast::TerminalExpr>(std::move(value));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<call_arg> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto arg = popExpr();
-            active_args.push_back(std::move(arg));
-        }
-    };
-
-    template <> struct action<call> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto callee = popIdentifier();
-            auto expr = std::make_unique<ast::CallExpr>(std::move(callee),
-                                                        std::move(active_args));
-            active_args.clear();
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<parens> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto back = popExpr();
-            auto expr = std::make_unique<ast::ParenExpr>(std::move(back));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<unary_plus> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto back = popExpr();
-            auto expr = std::make_unique<ast::UnaryOpExpr>(UnaryOp::PLUS,
-                                                           std::move(back));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<unary_minus> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto back = popExpr();
-            auto expr = std::make_unique<ast::UnaryOpExpr>(UnaryOp::MINUS,
-                                                           std::move(back));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<unary_bitwise_not> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto back = popExpr();
-            auto expr = std::make_unique<ast::UnaryOpExpr>(UnaryOp::NOT,
-                                                           std::move(back));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<deref> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto back = popExpr();
-            auto expr = std::make_unique<ast::UnaryOpExpr>(UnaryOp::DEREF,
-                                                           std::move(back));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<address> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto back = popExpr();
-            auto expr = std::make_unique<ast::UnaryOpExpr>(UnaryOp::ADDRESS,
-                                                           std::move(back));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<multiply> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto right = popExpr();
-            auto left = popExpr();
-            auto expr = std::make_unique<ast::BinaryOpExpr>(
-                BinaryOp::MUL, std::move(left), std::move(right));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<divide> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto right = popExpr();
-            auto left = popExpr();
-            auto expr = std::make_unique<ast::BinaryOpExpr>(
-                BinaryOp::DIV, std::move(left), std::move(right));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<add> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto right = popExpr();
-            auto left = popExpr();
-            auto expr = std::make_unique<ast::BinaryOpExpr>(
-                BinaryOp::ADD, std::move(left), std::move(right));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<subtract> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto right = popExpr();
-            auto left = popExpr();
-            auto expr = std::make_unique<ast::BinaryOpExpr>(
-                BinaryOp::SUB, std::move(left), std::move(right));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<left_shift> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto right = popExpr();
-            auto left = popExpr();
-            auto expr = std::make_unique<ast::BinaryOpExpr>(
-                BinaryOp::LSHIFT, std::move(left), std::move(right));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<right_shift> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto right = popExpr();
-            auto left = popExpr();
-            auto expr = std::make_unique<ast::BinaryOpExpr>(
-                BinaryOp::RSHIFT, std::move(left), std::move(right));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<bitwise_and> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto right = popExpr();
-            auto left = popExpr();
-            auto expr = std::make_unique<ast::BinaryOpExpr>(
-                BinaryOp::AND, std::move(left), std::move(right));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<bitwise_xor> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto right = popExpr();
-            auto left = popExpr();
-            auto expr = std::make_unique<ast::BinaryOpExpr>(
-                BinaryOp::XOR, std::move(left), std::move(right));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<bitwise_or> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto right = popExpr();
-            auto left = popExpr();
-            auto expr = std::make_unique<ast::BinaryOpExpr>(
-                BinaryOp::OR, std::move(left), std::move(right));
-            parsed_exprs.push_back(std::move(expr));
-        }
-    };
-
-    template <> struct action<type_int> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto type = std::make_unique<Int>();
-            parsed_types.push_back(std::move(type));
-        }
-    };
-
-    template <> struct action<type_ptr> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto base = popType();
-            auto type = std::make_unique<Ptr>(std::move(base));
-            parsed_types.push_back(std::move(type));
-        }
-    };
-
-    template <> struct action<add_equals> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            last_op_equals = BinaryOp::ADD;
-        }
-    };
-
-    template <> struct action<sub_equals> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            last_op_equals = BinaryOp::SUB;
-        }
-    };
-
-    template <> struct action<mul_equals> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            last_op_equals = BinaryOp::MUL;
-        }
-    };
-
-    template <> struct action<div_equals> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            last_op_equals = BinaryOp::DIV;
-        }
-    };
-
-    template <> struct action<lshift_equals> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            last_op_equals = BinaryOp::LSHIFT;
-        }
-    };
-
-    template <> struct action<rshift_equals> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            last_op_equals = BinaryOp::RSHIFT;
-        }
-    };
-
-    template <> struct action<and_equals> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            last_op_equals = BinaryOp::AND;
-        }
-    };
-
-    template <> struct action<or_equals> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            last_op_equals = BinaryOp::OR;
-        }
-    };
-
-    template <> struct action<xor_equals> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            last_op_equals = BinaryOp::XOR;
-        }
-    };
-
-    // Instruction actions
-    template <> struct action<left_brace> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto scope = std::make_unique<ast::Scope>();
-            active_scopes.push_back(std::move(scope));
-        }
-    };
-
-    template <> struct action<right_brace> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            if (active_scopes.size() == 1) // function -- in global scope
-                return;
-
-            auto scope = std::move(active_scopes.back());
-            active_scopes.pop_back();
-            active_scopes.back()->addInstruction(std::move(scope));
-        }
-    };
-
-    template <> struct action<instruction_expr> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto expr = popExpr();
-
-            auto i = std::make_unique<ast::InstructionExpr>(std::move(expr));
-            active_scopes.back()->addInstruction(std::move(i));
-        }
-    };
-
-    template <> struct action<instruction_declaration> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto variable = popIdentifier();
-            auto type = popType();
-
-            auto i = std::make_unique<ast::InstructionDeclaration>(
-                std::move(type), std::move(variable));
-            active_scopes.back()->addInstruction(std::move(i));
-        }
-    };
-
-    template <> struct action<instruction_declaration_assign> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto value = popExpr();
-            auto variable = popIdentifier();
-            auto type = popType();
-
-            auto i = std::make_unique<ast::InstructionDeclarationAssign>(
-                std::move(type), std::move(variable), std::move(value));
-            active_scopes.back()->addInstruction(std::move(i));
-        }
-    };
-
-    template <> struct action<instruction_assign> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto value = popExpr();
-            auto variable = popIdentifier();
-
-            auto i = std::make_unique<ast::InstructionAssign>(
-                std::move(variable), std::move(value));
-            active_scopes.back()->addInstruction(std::move(i));
-        }
-    };
-
-    template <> struct action<instruction_op_assign> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto value = popExpr();
-            auto variable = popIdentifier();
-
-            auto i = std::make_unique<ast::InstructionOpAssign>(
-                std::move(variable), last_op_equals, std::move(value));
-            active_scopes.back()->addInstruction(std::move(i));
-        }
-    };
-
-    template <> struct action<instruction_return> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto value = popExpr();
-
-            auto i = std::make_unique<ast::InstructionReturn>(std::move(value));
-            active_scopes.back()->addInstruction(std::move(i));
-        }
-    };
-
-    template <> struct action<instruction_if> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto t_branch = popInstruction();
-            auto cond = popExpr();
-
-            auto i = std::make_unique<ast::InstructionIf>(
-                std::move(cond), std::move(t_branch), nullptr);
-            active_scopes.back()->addInstruction(std::move(i));
-        }
-    };
-
-    template <> struct action<instruction_if_else> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto f_branch = popInstruction();
-            auto t_branch = popInstruction();
-            auto cond = popExpr();
-
-            auto i = std::make_unique<ast::InstructionIf>(
-                std::move(cond), std::move(t_branch), std::move(f_branch));
-            active_scopes.back()->addInstruction(std::move(i));
-        }
-    };
-
-    template <> struct action<instruction_while> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto body = popInstruction();
-            auto cond = popExpr();
-
-            auto i = std::make_unique<ast::InstructionWhile>(std::move(cond),
-                                                             std::move(body));
-            active_scopes.back()->addInstruction(std::move(i));
-        }
-    };
-
-    // Function
-    template <> struct action<function_param> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto param_name = popIdentifier();
-            auto param_type = popType();
-            active_params.push_back(
-                {std::move(param_type), std::move(param_name)});
-        }
-    };
-
-    template <> struct action<function> {
-        template <typename Input>
-        static void apply(const Input &in, std::vector<ast::Function> &res) {
-            auto body = popScope();
-
-            auto name = popIdentifier();
-            auto type = popType();
-
-            res.push_back(ast::Function(std::move(type), std::move(name),
-                                        std::move(active_params),
-                                        std::move(body)));
-            active_params.clear();
-        }
-    };
-
-    //
-    // Parse function
-    //
     ast::Program parse(const std::string &input_file) {
         file_input in(input_file);
