+++ src/frontend/parser/core.cpp
@@ -1,0 +1,5 @@
+#include "frontend/parser/core.h"
+
+namespace frontend {
+    std::unique_ptr<SymbolTable> symbol_table;
+}
