+++ src/frontend/parser/function.cpp
@@ -1,0 +1,5 @@
+#include "frontend/parser/function.h"
+
+namespace frontend {
+    std::vector<ast::Parameter> active_params;
+}
